[FIX] 뷰에서 (사용자 화면이라고 예를 들면) 입력한 값을 전달하는 값 객체 이름 변경
